+++ src/websockets/legacy/client.py
@@ -391,7 +391,7 @@ class WebSocketClientProtocol(WebSocketCommonProtocol):
                 raise InvalidHeader("Location")
             raise RedirectHandshake(response_headers["Location"])
         elif status_code != 101:
-            raise InvalidStatusCode(status_code)
+            raise InvalidStatusCode(status_code, response_headers)
 
         check_response(response_headers, key)